+++ packages/plugins/auto-doc/src/utils/pdfGenerator.ts
@@ -28,7 +28,9 @@ async function generatePdf(templateTitle: string, allBlocks: Element[]) {
 		text: handleRichTextEditorBlock,
 		image: processImageForPdfGeneration,
 		signalList: processSignalListForPdfGeneration,
-		table: processTableForPdfGeneration
+		table: processTableForPdfGeneration,
+		network: processImageForPdfGeneration,
+		communication: processImageForPdfGeneration
 	}
 
 	function incrementVerticalPositionForNextLine(lineHeight?: number) {